ендпоинты для вывода

+++ migrations/000006_create_orders_table.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS orders (
     user_id    BIGINT       NOT NULL REFERENCES users (id) ON DELETE CASCADE,
     order_id   VARCHAR(48)  NOT NULL UNIQUE,
     status     order_status NOT NULL DEFAULT 'NEW',
-    accrual    INT,
+    accrual    NUMERIC(20, 2),
     created_at TIMESTAMPTZ  NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMPTZ  NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
